Move OAuth token processing to app entry point
- Create oauth-interceptor.ts that runs BEFORE main.tsx
- Only runs on /auth/callback route (checks window.location.pathname)
- Processes tokens before PGLite DB initialization starts
- Removes module-level code from callback.tsx route file

This fixes the issue where callback.tsx module was loading on all pages
and OAuth tokens weren't being intercepted before app initialization.
